rename to indexer_types.h
